requirements.txt
anyio==3.6.1
asgiref==3.5.2
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.3
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.77.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
jose==1.0.0
msgpack==1.0.4
numpy==1.22.4
pandas==1.4.2
passlib==1.7.4
py-algorand-sdk==1.15.0
pyasn1==0.4.8
pycparser==2.21
pycrypto==2.6.1
pycryptodomex==3.14.1
pydantic==1.9.1
pymongo==4.1.1
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
requests==2.28.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.20.3
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
